+++ Modules/Shop/Routes/api.php
@@ -18,5 +18,12 @@ Route::middleware(['auth', 'language'])
     ->name('shop.')
     ->namespace('API')
     ->group(function(){
-        // ...
+        Route::middleware(['can:validate-shop-coupons'])->group(function() {
+            Route::get('/', 'ShopController@listCards')->name('listCards');
+            Route::get('card', 'ShopController@getCard')->name('getCard');
+            Route::patch('card/{handout}/redeem', 'ShopController@redeemCard')->name('redeemCard');
+            Route::delete('card/{handout}', 'ShopController@cancelCard')->name('cancelCard');
+            Route::get('/summary', 'ShopController@summary')->name('summary');
+            Route::post('/deleteNonRedeemed', 'ShopController@deleteNonRedeemed')->name('deleteNonRedeemed');
+        });
     });