+++ resources/js/utils.js
@@ -7,7 +7,7 @@ export function showSnackbar(text, actionText, actionClass, callback) {
 		pos: 'bottom-center',
 		actionText: actionText ? actionText : null,
 		actionTextColor: null,
-		customClass: actionClass ? actionClass : null, 
+		customClass: actionClass ? actionClass : null,
 	};
 	if (callback) {
 		args['onActionClick'] = callback;
@@ -28,6 +28,9 @@ export function getAjaxErrorMessage(err) {
 			});
 		} else if (err.response.data.error) {
 			msg = err.response.data.error;
+		}
+		if (!msg) {
+			msg = `Error ${err.response.status}: ${err.response.statusText}`
 		}
 	} else {
 		msg = err